updates after code review
