+++ app-libs/form-component/src/layout-components/DemoLayoutComponent/DemoLayoutComponent.mdx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/addon-docs/blocks';
 
 import { LayoutComponentDocs } from '../common/storybook';
 import * as DemoLayoutComponentStories from './DemoLayoutComponent.stories';
-import { DEMO_PROP_CATEGORIES } from './DemoLayoutComponent';
+import { DEMO_PROP_CATEGORIES } from './DemoLayoutComponent.stories';
 
 <Meta of={DemoLayoutComponentStories} />
